itip_support.c: move sched_deliver_local() and assorted support functions from http_caldav_sched.c to itip_support.c
